Chrisotpher Lozinski - 2024-04-09
Here is an updated version of the Pico-Ice port.
The big change is that is that the ICE40UP5K has 30 separate four kbit ram blocks, not 32. So that is now fixed in both the verilog and the Forth.
I also moved the verilog from ./icestorm to ./verilog.
ALso generally I am just much better at workingj with this code.
It is a git repository, so that it is easy to see the changes.
I am busy trying to test it with e4thcom, so that is why the errors were caught.

Info: Device utilisation:
Info: ICESTORM_LC: 2787/ 5280 52%
Info: ICESTORM_RAM: 30/ 30 100%
Download pico-ice2.6d.tar
tar xvf pico-ice2.6d.tar